Extract estimated coefficients

Extracts estimated coefficients from linear and logistic regression, under the penalty parameter that minimises the cross-validated loss.

# S3 method for cornet
coef(object, ...)

Arguments

object

cornet object

...

further arguments (not applicable)

Value

This function returns a matrix with \(n\) rows and two columns, where \(n\) is the sample size. It includes the estimated coefficients from linear regression (1st column: "beta") and logistic regression (2nd column: "gamma").

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
coef(net)
#> 201 x 2 sparse Matrix of class "dgCMatrix"
#>                      beta     gamma
#> (Intercept)  4.323128e-02 0.2818512
#> V1           .            .        
#> V2           .            .        
#> V3           .            .        
#> V4           .            .        
#> V5           .            .        
#> V6           .            .        
#> V7           .            .        
#> V8           .            .        
#> V9           .            .        
#> V10          8.110151e-02 .        
#> V11          1.836699e-04 .        
#> V12          .            .        
#> V13          .            .        
#> V14          .            .        
#> V15          .            .        
#> V16          .            .        
#> V17          .            .        
#> V18          .            .        
#> V19          .            .        
#> V20          .            .        
#> V21          .            .        
#> V22          .            .        
#> V23          .            .        
#> V24          .            .        
#> V25          .            .        
#> V26          .            .        
#> V27          .            .        
#> V28          .            .        
#> V29          .            .        
#> V30          .            .        
#> V31          .            .        
#> V32          .            .        
#> V33          .            .        
#> V34          .            .        
#> V35          .            .        
#> V36          .            .        
#> V37          .            .        
#> V38          .            .        
#> V39          .            .        
#> V40          .            .        
#> V41          .            .        
#> V42          .            .        
#> V43          .            .        
#> V44          .            .        
#> V45          .            .        
#> V46          .            .        
#> V47          .            .        
#> V48          .            .        
#> V49          .            .        
#> V50          .            .        
#> V51          .            .        
#> V52          .            .        
#> V53          4.522114e-02 .        
#> V54          .            .        
#> V55          .            .        
#> V56          .            .        
#> V57         -3.628373e-03 .        
#> V58          .            .        
#> V59          .            .        
#> V60          .            .        
#> V61          .            .        
#> V62          .            .        
#> V63         -4.088288e-02 .        
#> V64          .            .        
#> V65          .            .        
#> V66          .            .        
#> V67          .            .        
#> V68          .            .        
#> V69          .            .        
#> V70          .            .        
#> V71          .            .        
#> V72          .            .        
#> V73          .            .        
#> V74          .            .        
#> V75          .            .        
#> V76          .            .        
#> V77          .            .        
#> V78          .            .        
#> V79          .            .        
#> V80          .            .        
#> V81          .            .        
#> V82          .            .        
#> V83          .            .        
#> V84          .            .        
#> V85          .            .        
#> V86          .            .        
#> V87          .            .        
#> V88          .            .        
#> V89          .            .        
#> V90          .            .        
#> V91          .            .        
#> V92          .            .        
#> V93          .            .        
#> V94          .            .        
#> V95          .            .        
#> V96          .            .        
#> V97          .            .        
#> V98         -3.705848e-02 .        
#> V99          .            .        
#> V100         .            .        
#> V101         .            .        
#> V102         .            .        
#> V103         .            .        
#> V104         .            .        
#> V105         .            .        
#> V106         .            .        
#> V107         .            .        
#> V108         .            .        
#> V109         .            .        
#> V110         .            .        
#> V111         .            .        
#> V112         .            .        
#> V113         .            .        
#> V114         .            .        
#> V115         .            .        
#> V116         3.476956e-02 .        
#> V117         .            .        
#> V118         .            .        
#> V119         .            .        
#> V120        -1.856054e-02 .        
#> V121         .            .        
#> V122         .            .        
#> V123         .            .        
#> V124         .            .        
#> V125         .            .        
#> V126         .            .        
#> V127         .            .        
#> V128         .            .        
#> V129         .            .        
#> V130         .            .        
#> V131         .            .        
#> V132         .            .        
#> V133         .            .        
#> V134         .            .        
#> V135         .            .        
#> V136         .            .        
#> V137         .            .        
#> V138         .            .        
#> V139         .            .        
#> V140         .            .        
#> V141         .            .        
#> V142         .            .        
#> V143         .            .        
#> V144        -6.370954e-03 .        
#> V145         1.280135e-01 .        
#> V146         .            .        
#> V147         .            .        
#> V148         .            .        
#> V149        -8.293407e-02 .        
#> V150         .            .        
#> V151         .            .        
#> V152         .            .        
#> V153         .            .        
#> V154         .            .        
#> V155         .            .        
#> V156         .            .        
#> V157         .            .        
#> V158         .            .        
#> V159         .            .        
#> V160         .            .        
#> V161         .            .        
#> V162         .            .        
#> V163         .            .        
#> V164         2.249576e-02 .        
#> V165         .            .        
#> V166         .            .        
#> V167         .            .        
#> V168         .            .        
#> V169         .            .        
#> V170         .            .        
#> V171         .            .        
#> V172         .            .        
#> V173         .            .        
#> V174         .            .        
#> V175         .            .        
#> V176         .            .        
#> V177         .            .        
#> V178         .            .        
#> V179         .            .        
#> V180         .            .        
#> V181         .            .        
#> V182         .            .        
#> V183        -5.204400e-02 .        
#> V184        -9.893018e-06 .        
#> V185         .            .        
#> V186         .            .        
#> V187         .            .        
#> V188         .            .        
#> V189         .            .        
#> V190         .            .        
#> V191         .            .        
#> V192        -1.151237e-02 .        
#> V193         .            .        
#> V194         .            .        
#> V195         .            .        
#> V196         .            .        
#> V197         2.870475e-02 .        
#> V198         .            .        
#> V199         .            .        
#> V200         .            .