Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
#> Multiple minima.
predict(net,newx=X)
#>     intercept  gaussian  binomial  combined    switch
#> 1        0.54 0.5666180 0.4073551 0.4238353 0.4073551
#> 2        0.54 0.5892086 0.5561136 0.5636810 0.5561136
#> 3        0.54 0.5971813 0.6704113 0.6698723 0.6704113
#> 4        0.54 0.5700342 0.5279073 0.5351992 0.5279073
#> 5        0.54 0.5897630 0.7308523 0.7245130 0.7308523
#> 6        0.54 0.5708044 0.4589926 0.4719004 0.4589926
#> 7        0.54 0.5513678 0.4607731 0.4709297 0.4607731
#> 8        0.54 0.5784523 0.5104279 0.5202358 0.5104279
#> 9        0.54 0.5803896 0.5062790 0.5166746 0.5062790
#> 10       0.54 0.5882170 0.5896927 0.5944440 0.5896927
#> 11       0.54 0.5829031 0.4900484 0.5020738 0.4900484
#> 12       0.54 0.5951459 0.5567549 0.5650445 0.5567549
#> 13       0.54 0.6029691 0.6624875 0.6633291 0.6624875
#> 14       0.54 0.5719748 0.4350663 0.4500439 0.4350663
#> 15       0.54 0.5934133 0.6367835 0.6384456 0.6367835
#> 16       0.54 0.5850172 0.6166159 0.6187938 0.6166159
#> 17       0.54 0.5908797 0.6704993 0.6691341 0.6704993
#> 18       0.54 0.5708606 0.3860241 0.4047768 0.3860241
#> 19       0.54 0.5810909 0.4591022 0.4733645 0.4591022
#> 20       0.54 0.5675710 0.5037251 0.5126231 0.5037251
#> 21       0.54 0.5885738 0.6532621 0.6529745 0.6532621
#> 22       0.54 0.5924333 0.5879935 0.5934313 0.5879935
#> 23       0.54 0.5876273 0.5527932 0.5604193 0.5527932
#> 24       0.54 0.5852616 0.6491301 0.6487390 0.6491301
#> 25       0.54 0.5615729 0.4457101 0.4584462 0.4457101
#> 26       0.54 0.5639689 0.4874985 0.4972128 0.4874985
#> 27       0.54 0.5953614 0.5915551 0.5970886 0.5915551
#> 28       0.54 0.5466931 0.5166560 0.5217092 0.5166560
#> 29       0.54 0.5798882 0.4844656 0.4965401 0.4844656
#> 30       0.54 0.5944770 0.6579471 0.6580544 0.6579471
#> 31       0.54 0.5706010 0.5294341 0.5366794 0.5294341
#> 32       0.54 0.5903259 0.5130704 0.5242271 0.5130704
#> 33       0.54 0.5807678 0.5809343 0.5854074 0.5809343
#> 34       0.54 0.5683672 0.4370668 0.4514036 0.4370668
#> 35       0.54 0.5714068 0.4422640 0.4565902 0.4422640
#> 36       0.54 0.5822384 0.4194661 0.4370505 0.4194661
#> 37       0.54 0.5576855 0.5283045 0.5339107 0.5283045
#> 38       0.54 0.5800787 0.5168481 0.5263572 0.5168481
#> 39       0.54 0.5826954 0.5337689 0.5422693 0.5337689
#> 40       0.54 0.5640302 0.4401104 0.4536239 0.4401104
#> 41       0.54 0.5879297 0.5353789 0.5444377 0.5353789
#> 42       0.54 0.5694325 0.6409204 0.6390911 0.6409204
#> 43       0.54 0.5915331 0.6305660 0.6324807 0.6305660
#> 44       0.54 0.5843435 0.4917575 0.5038356 0.4917575
#> 45       0.54 0.5678814 0.4174441 0.4332859 0.4174441
#> 46       0.54 0.5720635 0.4331983 0.4483371 0.4331983
#> 47       0.54 0.5672147 0.5342893 0.5406944 0.5342893
#> 48       0.54 0.5711193 0.4961991 0.5061722 0.4961991
#> 49       0.54 0.5941607 0.5743087 0.5810660 0.5743087
#> 50       0.54 0.5581123 0.4708731 0.4811311 0.4708731
#> 51       0.54 0.5841432 0.5485404 0.5560496 0.5485404
#> 52       0.54 0.5794522 0.5406629 0.5481841 0.5406629
#> 53       0.54 0.5743873 0.4855126 0.4967751 0.4855126
#> 54       0.54 0.5786280 0.5303985 0.5386319 0.5303985
#> 55       0.54 0.5773213 0.5712946 0.5760835 0.5712946
#> 56       0.54 0.5641621 0.5398752 0.5454252 0.5398752
#> 57       0.54 0.5520999 0.4221716 0.4355152 0.4221716
#> 58       0.54 0.5780203 0.5035698 0.5138692 0.5035698
#> 59       0.54 0.5719955 0.5660152 0.5705196 0.5660152
#> 60       0.54 0.5767823 0.4815424 0.4934401 0.4815424
#> 61       0.54 0.5817379 0.5628054 0.5688568 0.5628054
#> 62       0.54 0.5777407 0.6348680 0.6346265 0.6348680
#> 63       0.54 0.5605346 0.4016806 0.4177997 0.4016806
#> 64       0.54 0.5864408 0.4127306 0.4314062 0.4127306
#> 65       0.54 0.5876256 0.6286571 0.6302138 0.6286571
#> 66       0.54 0.6055322 0.6856818 0.6849968 0.6856818
#> 67       0.54 0.5757272 0.5634229 0.5686303 0.5634229
#> 68       0.54 0.5458674 0.3970722 0.4115802 0.3970722
#> 69       0.54 0.5865528 0.6327058 0.6337981 0.6327058
#> 70       0.54 0.5879372 0.5779901 0.5836410 0.5779901
#> 71       0.54 0.5610897 0.4689373 0.4797504 0.4689373
#> 72       0.54 0.5763916 0.6718143 0.6684385 0.6718143
#> 73       0.54 0.5890202 0.5609041 0.5680636 0.5609041
#> 74       0.54 0.5920092 0.6051342 0.6091455 0.6051342
#> 75       0.54 0.5841251 0.6512128 0.6505058 0.6512128
#> 76       0.54 0.5847645 0.5019473 0.5132655 0.5019473
#> 77       0.54 0.5789891 0.6205069 0.6215794 0.6205069
#> 78       0.54 0.5625447 0.4182161 0.4332821 0.4182161
#> 79       0.54 0.5607407 0.4606715 0.4720990 0.4606715
#> 80       0.54 0.5895522 0.5624786 0.5695817 0.5624786
#> 81       0.54 0.5775571 0.5270279 0.5353893 0.5270279
#> 82       0.54 0.5661994 0.4401296 0.4539318 0.4401296
#> 83       0.54 0.5757357 0.6201226 0.6207952 0.6201226
#> 84       0.54 0.5735836 0.5113267 0.5204173 0.5113267
#> 85       0.54 0.5888327 0.6123204 0.6153420 0.6123204
#> 86       0.54 0.5726017 0.4296883 0.4451796 0.4296883
#> 87       0.54 0.5815632 0.5466370 0.5539588 0.5466370
#> 88       0.54 0.5923819 0.5843825 0.5901024 0.5843825
#> 89       0.54 0.5979940 0.6757032 0.6748460 0.6757032
#> 90       0.54 0.5727079 0.4246452 0.4405540 0.4246452
#> 91       0.54 0.5887687 0.5710982 0.5774093 0.5710982
#> 92       0.54 0.5684652 0.5514788 0.5566758 0.5514788
#> 93       0.54 0.5678882 0.5130323 0.5212279 0.5130323
#> 94       0.54 0.5924303 0.6244063 0.6269307 0.6244063
#> 95       0.54 0.5771391 0.5929300 0.5959639 0.5929300
#> 96       0.54 0.5723778 0.6151877 0.6158092 0.6151877
#> 97       0.54 0.5841176 0.6171407 0.6191584 0.6171407
#> 98       0.54 0.5526932 0.4993831 0.5066299 0.4993831
#> 99       0.54 0.5793148 0.5382087 0.5459080 0.5382087
#> 100      0.54 0.5884048 0.6670032 0.6655942 0.6670032