Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)
#>     intercept  gaussian binomial  combined    switch
#> 1        0.52 0.4364821     0.52 0.4298895 0.4364821
#> 2        0.52 0.6189245     0.52 0.6515215 0.6189245
#> 3        0.52 0.4731316     0.52 0.4755177 0.4731316
#> 4        0.52 0.4873237     0.52 0.4937795 0.4873237
#> 5        0.52 0.4899237     0.52 0.4971425 0.4899237
#> 6        0.52 0.4306019     0.52 0.4228799 0.4306019
#> 7        0.52 0.5885479     0.52 0.6195510 0.5885479
#> 8        0.52 0.4247404     0.52 0.4160013 0.4247404
#> 9        0.52 0.6835931     0.52 0.7047018 0.6835931
#> 10       0.52 0.5362868     0.52 0.5568653 0.5362868
#> 11       0.52 0.5123820     0.52 0.5262401 0.5123820
#> 12       0.52 0.6526542     0.52 0.6819542 0.6526542
#> 13       0.52 0.4200526     0.52 0.4105833 0.4200526
#> 14       0.52 0.7216736     0.52 0.7258451 0.7216736
#> 15       0.52 0.4294184     0.52 0.4214821 0.4294184
#> 16       0.52 0.5734912     0.52 0.6023354 0.5734912
#> 17       0.52 0.5764370     0.52 0.6057667 0.5764370
#> 18       0.52 0.3945144     0.52 0.3825303 0.3945144
#> 19       0.52 0.4817430     0.52 0.4865760 0.4817430
#> 20       0.52 0.7391511     0.52 0.7331576 0.7391511
#> 21       0.52 0.4854003     0.52 0.4912942 0.4854003
#> 22       0.52 0.5220313     0.52 0.5386806 0.5220313
#> 23       0.52 0.5808129     0.52 0.6108086 0.5808129
#> 24       0.52 0.4754046     0.52 0.4784285 0.4754046
#> 25       0.52 0.4911016     0.52 0.4986674 0.4911016
#> 26       0.52 0.5597866     0.52 0.5860095 0.5597866
#> 27       0.52 0.5038976     0.52 0.5152535 0.5038976
#> 28       0.52 0.4956746     0.52 0.5045920 0.4956746
#> 29       0.52 0.5703363     0.52 0.5986286 0.5703363
#> 30       0.52 0.5109842     0.52 0.5244323 0.5109842
#> 31       0.52 0.5098187     0.52 0.5229241 0.5098187
#> 32       0.52 0.4125374     0.52 0.4020622 0.4125374
#> 33       0.52 0.6960280     0.52 0.7124235 0.6960280
#> 34       0.52 0.6023386     0.52 0.6345566 0.6023386
#> 35       0.52 0.4237108     0.52 0.4148048 0.4237108
#> 36       0.52 0.4592916     0.52 0.4579600 0.4592916
#> 37       0.52 0.6032662     0.52 0.6355376 0.6032662
#> 38       0.52 0.6415247     0.52 0.6725451 0.6415247
#> 39       0.52 0.4578115     0.52 0.4561022 0.4578115
#> 40       0.52 0.4039370     0.52 0.3925757 0.4039370
#> 41       0.52 0.5766042     0.52 0.6059605 0.5766042
#> 42       0.52 0.5043036     0.52 0.5157797 0.5043036
#> 43       0.52 0.4215856     0.52 0.4123467 0.4215856
#> 44       0.52 0.4174547     0.52 0.4076142 0.4174547
#> 45       0.52 0.5512080     0.52 0.5755187 0.5512080
#> 46       0.52 0.5251605     0.52 0.5426958 0.5251605
#> 47       0.52 0.6268205     0.52 0.6591508 0.6268205
#> 48       0.52 0.3698974     0.52 0.3581665 0.3698974
#> 49       0.52 0.2418980     0.52 0.2808525 0.2418980
#> 50       0.52 0.6175271     0.52 0.6501405 0.6175271
#> 51       0.52 0.6176426     0.52 0.6502549 0.6176426
#> 52       0.52 0.4695466     0.52 0.4709409 0.4695466
#> 53       0.52 0.4358359     0.52 0.4291141 0.4358359
#> 54       0.52 0.5807360     0.52 0.6107205 0.5807360
#> 55       0.52 0.4549873     0.52 0.4525696 0.4549873
#> 56       0.52 0.5977533     0.52 0.6296539 0.5977533
#> 57       0.52 0.5122777     0.52 0.5261052 0.5122777
#> 58       0.52 0.5197183     0.52 0.5357060 0.5197183
#> 59       0.52 0.4330677     0.52 0.4258064 0.4330677
#> 60       0.52 0.4493701     0.52 0.4455950 0.4493701
#> 61       0.52 0.4791709     0.52 0.4832648 0.4791709
#> 62       0.52 0.4944436     0.52 0.5029965 0.4944436
#> 63       0.52 0.3288769     0.52 0.3242586 0.3288769
#> 64       0.52 0.4985142     0.52 0.5082734 0.4985142
#> 65       0.52 0.5945111     0.52 0.6261343 0.5945111
#> 66       0.52 0.6480268     0.52 0.6781197 0.6480268
#> 67       0.52 0.5728276     0.52 0.6015583 0.5728276
#> 68       0.52 0.5102528     0.52 0.5234859 0.5102528
#> 69       0.52 0.3867036     0.52 0.3744959 0.3867036
#> 70       0.52 0.6365006     0.52 0.6680900 0.6365006
#> 71       0.52 0.4520708     0.52 0.4489394 0.4520708
#> 72       0.52 0.5999256     0.52 0.6319876 0.5999256
#> 73       0.52 0.4263512     0.52 0.4178803 0.4263512
#> 74       0.52 0.4505214     0.52 0.4470187 0.4505214
#> 75       0.52 0.5934888     0.52 0.6250157 0.5934888
#> 76       0.52 0.5357710     0.52 0.5562126 0.5357710
#> 77       0.52 0.4401310     0.52 0.4342907 0.4401310
#> 78       0.52 0.5509449     0.52 0.5751939 0.5509449
#> 79       0.52 0.4458931     0.52 0.4413151 0.4458931
#> 80       0.52 0.5182060     0.52 0.5337584 0.5182060
#> 81       0.52 0.4698522     0.52 0.4713303 0.4698522
#> 82       0.52 0.5617766     0.52 0.5884150 0.5617766
#> 83       0.52 0.5869454     0.52 0.6177582 0.5869454
#> 84       0.52 0.5325911     0.52 0.5521795 0.5325911
#> 85       0.52 0.5144100     0.52 0.5288609 0.5144100
#> 86       0.52 0.4379458     0.52 0.4316503 0.4379458
#> 87       0.52 0.2669826     0.52 0.2896471 0.2669826
#> 88       0.52 0.4482067     0.52 0.4441597 0.4482067
#> 89       0.52 0.3032824     0.52 0.3075719 0.3032824
#> 90       0.52 0.5677242     0.52 0.5955355 0.5677242
#> 91       0.52 0.6413845     0.52 0.6724225 0.6413845
#> 92       0.52 0.6779881     0.52 0.7009557 0.6779881
#> 93       0.52 0.5246957     0.52 0.5421001 0.5246957
#> 94       0.52 0.5681043     0.52 0.5959869 0.5681043
#> 95       0.52 0.7304705     0.52 0.7297037 0.7304705
#> 96       0.52 0.5773164     0.52 0.6067853 0.5773164
#> 97       0.52 0.3116501     0.52 0.3126488 0.3116501
#> 98       0.52 0.4521671     0.52 0.4490589 0.4521671
#> 99       0.52 0.5377211     0.52 0.5586776 0.5377211
#> 100      0.52 0.5051984     0.52 0.5169396 0.5051984