Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
#> Multiple minima.
#> Multiple minima.
predict(net,newx=X)
#>     intercept  gaussian  binomial  combined    switch
#> 1        0.47 0.4963683 0.4954927 0.4954927 0.4954927
#> 2        0.47 0.4963683 0.4336249 0.4336249 0.4336249
#> 3        0.47 0.4963683 0.4245905 0.4245905 0.4245905
#> 4        0.47 0.4963683 0.4703704 0.4703704 0.4703704
#> 5        0.47 0.4963683 0.4638743 0.4638743 0.4638743
#> 6        0.47 0.4963683 0.4423949 0.4423949 0.4423949
#> 7        0.47 0.4963683 0.4997668 0.4997668 0.4997668
#> 8        0.47 0.4963683 0.4836120 0.4836120 0.4836120
#> 9        0.47 0.4963683 0.4581131 0.4581131 0.4581131
#> 10       0.47 0.4963683 0.5162469 0.5162469 0.5162469
#> 11       0.47 0.4963683 0.4715239 0.4715239 0.4715239
#> 12       0.47 0.4963683 0.5363589 0.5363589 0.5363589
#> 13       0.47 0.4963683 0.4368404 0.4368404 0.4368404
#> 14       0.47 0.4963683 0.4354912 0.4354912 0.4354912
#> 15       0.47 0.4963683 0.3997009 0.3997009 0.3997009
#> 16       0.47 0.4963683 0.5039856 0.5039856 0.5039856
#> 17       0.47 0.4963683 0.4807287 0.4807287 0.4807287
#> 18       0.47 0.4963683 0.4737318 0.4737318 0.4737318
#> 19       0.47 0.4963683 0.4587870 0.4587870 0.4587870
#> 20       0.47 0.4963683 0.4308635 0.4308635 0.4308635
#> 21       0.47 0.4963683 0.4306178 0.4306178 0.4306178
#> 22       0.47 0.4963683 0.4393233 0.4393233 0.4393233
#> 23       0.47 0.4963683 0.4864562 0.4864562 0.4864562
#> 24       0.47 0.4963683 0.4686864 0.4686864 0.4686864
#> 25       0.47 0.4963683 0.4884268 0.4884268 0.4884268
#> 26       0.47 0.4963683 0.4999953 0.4999953 0.4999953
#> 27       0.47 0.4963683 0.5176204 0.5176204 0.5176204
#> 28       0.47 0.4963683 0.5202067 0.5202067 0.5202067
#> 29       0.47 0.4963683 0.4869127 0.4869127 0.4869127
#> 30       0.47 0.4963683 0.4959685 0.4959685 0.4959685
#> 31       0.47 0.4963683 0.4610039 0.4610039 0.4610039
#> 32       0.47 0.4963683 0.4791230 0.4791230 0.4791230
#> 33       0.47 0.4963683 0.4292204 0.4292204 0.4292204
#> 34       0.47 0.4963683 0.4511304 0.4511304 0.4511304
#> 35       0.47 0.4963683 0.4948851 0.4948851 0.4948851
#> 36       0.47 0.4963683 0.4802630 0.4802630 0.4802630
#> 37       0.47 0.4963683 0.4658293 0.4658293 0.4658293
#> 38       0.47 0.4963683 0.4622872 0.4622872 0.4622872
#> 39       0.47 0.4963683 0.3870489 0.3870489 0.3870489
#> 40       0.47 0.4963683 0.5373481 0.5373481 0.5373481
#> 41       0.47 0.4963683 0.4254143 0.4254143 0.4254143
#> 42       0.47 0.4963683 0.4968191 0.4968191 0.4968191
#> 43       0.47 0.4963683 0.4966946 0.4966946 0.4966946
#> 44       0.47 0.4963683 0.5977134 0.5977134 0.5977134
#> 45       0.47 0.4963683 0.3807017 0.3807017 0.3807017
#> 46       0.47 0.4963683 0.5067429 0.5067429 0.5067429
#> 47       0.47 0.4963683 0.4333289 0.4333289 0.4333289
#> 48       0.47 0.4963683 0.4932065 0.4932065 0.4932065
#> 49       0.47 0.4963683 0.4248429 0.4248429 0.4248429
#> 50       0.47 0.4963683 0.3891678 0.3891678 0.3891678
#> 51       0.47 0.4963683 0.4292093 0.4292093 0.4292093
#> 52       0.47 0.4963683 0.5829450 0.5829450 0.5829450
#> 53       0.47 0.4963683 0.4740985 0.4740985 0.4740985
#> 54       0.47 0.4963683 0.5139483 0.5139483 0.5139483
#> 55       0.47 0.4963683 0.4723041 0.4723041 0.4723041
#> 56       0.47 0.4963683 0.4943747 0.4943747 0.4943747
#> 57       0.47 0.4963683 0.4730588 0.4730588 0.4730588
#> 58       0.47 0.4963683 0.5624165 0.5624165 0.5624165
#> 59       0.47 0.4963683 0.4125187 0.4125187 0.4125187
#> 60       0.47 0.4963683 0.3759418 0.3759418 0.3759418
#> 61       0.47 0.4963683 0.4985044 0.4985044 0.4985044
#> 62       0.47 0.4963683 0.4564184 0.4564184 0.4564184
#> 63       0.47 0.4963683 0.5312626 0.5312626 0.5312626
#> 64       0.47 0.4963683 0.4225472 0.4225472 0.4225472
#> 65       0.47 0.4963683 0.3779018 0.3779018 0.3779018
#> 66       0.47 0.4963683 0.4411270 0.4411270 0.4411270
#> 67       0.47 0.4963683 0.5315449 0.5315449 0.5315449
#> 68       0.47 0.4963683 0.4575107 0.4575107 0.4575107
#> 69       0.47 0.4963683 0.5381645 0.5381645 0.5381645
#> 70       0.47 0.4963683 0.5457065 0.5457065 0.5457065
#> 71       0.47 0.4963683 0.5056756 0.5056756 0.5056756
#> 72       0.47 0.4963683 0.5052245 0.5052245 0.5052245
#> 73       0.47 0.4963683 0.5163180 0.5163180 0.5163180
#> 74       0.47 0.4963683 0.4190171 0.4190171 0.4190171
#> 75       0.47 0.4963683 0.5019739 0.5019739 0.5019739
#> 76       0.47 0.4963683 0.4331577 0.4331577 0.4331577
#> 77       0.47 0.4963683 0.4140789 0.4140789 0.4140789
#> 78       0.47 0.4963683 0.5390373 0.5390373 0.5390373
#> 79       0.47 0.4963683 0.3721754 0.3721754 0.3721754
#> 80       0.47 0.4963683 0.4660442 0.4660442 0.4660442
#> 81       0.47 0.4963683 0.5025104 0.5025104 0.5025104
#> 82       0.47 0.4963683 0.4405621 0.4405621 0.4405621
#> 83       0.47 0.4963683 0.4513145 0.4513145 0.4513145
#> 84       0.47 0.4963683 0.4980117 0.4980117 0.4980117
#> 85       0.47 0.4963683 0.4481357 0.4481357 0.4481357
#> 86       0.47 0.4963683 0.5089265 0.5089265 0.5089265
#> 87       0.47 0.4963683 0.4770842 0.4770842 0.4770842
#> 88       0.47 0.4963683 0.4682844 0.4682844 0.4682844
#> 89       0.47 0.4963683 0.4342081 0.4342081 0.4342081
#> 90       0.47 0.4963683 0.4046587 0.4046587 0.4046587
#> 91       0.47 0.4963683 0.4718238 0.4718238 0.4718238
#> 92       0.47 0.4963683 0.4710982 0.4710982 0.4710982
#> 93       0.47 0.4963683 0.4303556 0.4303556 0.4303556
#> 94       0.47 0.4963683 0.4686103 0.4686103 0.4686103
#> 95       0.47 0.4963683 0.4813132 0.4813132 0.4813132
#> 96       0.47 0.4963683 0.4710399 0.4710399 0.4710399
#> 97       0.47 0.4963683 0.4851774 0.4851774 0.4851774
#> 98       0.47 0.4963683 0.5004548 0.5004548 0.5004548
#> 99       0.47 0.4963683 0.4670165 0.4670165 0.4670165
#> 100      0.47 0.4963683 0.4121237 0.4121237 0.4121237