Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)
#>     intercept   binomial   combined
#> 1        0.44 0.22449314 0.19983538
#> 2        0.44 0.53401798 0.57360757
#> 3        0.44 0.79503195 0.70759082
#> 4        0.44 0.13339714 0.12129591
#> 5        0.44 0.85236619 0.86345321
#> 6        0.44 0.85559306 0.85629718
#> 7        0.44 0.83890693 0.83766484
#> 8        0.44 0.82337544 0.82189683
#> 9        0.44 0.65740816 0.59427493
#> 10       0.44 0.29025912 0.25839076
#> 11       0.44 0.66337788 0.67033519
#> 12       0.44 0.62155859 0.60220183
#> 13       0.44 0.24232179 0.31177059
#> 14       0.44 0.08169104 0.07273824
#> 15       0.44 0.74649597 0.68410067
#> 16       0.44 0.88213463 0.80819230
#> 17       0.44 0.08218402 0.07329611
#> 18       0.44 0.84202309 0.85760431
#> 19       0.44 0.10052849 0.08947049
#> 20       0.44 0.03814559 0.03401925
#> 21       0.44 0.17562729 0.15753870
#> 22       0.44 0.63090553 0.56150949
#> 23       0.44 0.91353161 0.90815292
#> 24       0.44 0.16642677 0.15223030
#> 25       0.44 0.88083713 0.85878590
#> 26       0.44 0.24104104 0.23675466
#> 27       0.44 0.13754616 0.12241904
#> 28       0.44 0.05122260 0.04585752
#> 29       0.44 0.80696097 0.72633643
#> 30       0.44 0.21160118 0.18833879
#> 31       0.44 0.73435604 0.70381643
#> 32       0.44 0.12261863 0.10913674
#> 33       0.44 0.27518208 0.30159005
#> 34       0.44 0.13030660 0.11598511
#> 35       0.44 0.20295106 0.18123984
#> 36       0.44 0.19627693 0.18861355
#> 37       0.44 0.10267048 0.09137701
#> 38       0.44 0.19910134 0.21750919
#> 39       0.44 0.69404967 0.63771227
#> 40       0.44 0.25833132 0.24862953
#> 41       0.44 0.86614247 0.82523457
#> 42       0.44 0.92164000 0.93025328
#> 43       0.44 0.04736826 0.04215816
#> 44       0.44 0.11460615 0.10214748
#> 45       0.44 0.04908110 0.04369960
#> 46       0.44 0.22988703 0.23327223
#> 47       0.44 0.23310405 0.27225406
#> 48       0.44 0.22335396 0.26373027
#> 49       0.44 0.07802376 0.06944117
#> 50       0.44 0.14648017 0.13039083
#> 51       0.44 0.65310506 0.63336354
#> 52       0.44 0.24225924 0.26958731
#> 53       0.44 0.13754473 0.12245314
#> 54       0.44 0.89045078 0.80155112
#> 55       0.44 0.89234172 0.89653185
#> 56       0.44 0.29334575 0.28968902
#> 57       0.44 0.70684765 0.62946133
#> 58       0.44 0.30586421 0.30401579
#> 59       0.44 0.12643941 0.11257952
#> 60       0.44 0.82708954 0.77609234
#> 61       0.44 0.89125090 0.79501500
#> 62       0.44 0.90690155 0.81478470
#> 63       0.44 0.72948720 0.71691901
#> 64       0.44 0.64493375 0.59938260
#> 65       0.44 0.69199092 0.65628110
#> 66       0.44 0.28194863 0.25093449
#> 67       0.44 0.87082342 0.78736983
#> 68       0.44 0.80418649 0.73882777
#> 69       0.44 0.16836982 0.15530012
#> 70       0.44 0.01370685 0.01219910
#> 71       0.44 0.35615595 0.40306095
#> 72       0.44 0.05857127 0.05212898
#> 73       0.44 0.16950755 0.15098005
#> 74       0.44 0.73508123 0.65510986
#> 75       0.44 0.68132603 0.62401863
#> 76       0.44 0.09112200 0.08266045
#> 77       0.44 0.05787371 0.06698394
#> 78       0.44 0.15123140 0.14115360
#> 79       0.44 0.90710740 0.81716701
#> 80       0.44 0.78174248 0.75401257
#> 81       0.44 0.03647565 0.06194622
#> 82       0.44 0.91405975 0.83034629
#> 83       0.44 0.25784415 0.24366295
#> 84       0.44 0.88677431 0.89859558
#> 85       0.44 0.79242711 0.80454960
#> 86       0.44 0.24404792 0.23214295
#> 87       0.44 0.09162274 0.11972540
#> 88       0.44 0.05445854 0.04988465
#> 89       0.44 0.35871391 0.32940585
#> 90       0.44 0.74724981 0.66507382
#> 91       0.44 0.93880897 0.90801515
#> 92       0.44 0.16088686 0.14320172
#> 93       0.44 0.17513834 0.15587337
#> 94       0.44 0.19437069 0.17309609
#> 95       0.44 0.14845045 0.20582280
#> 96       0.44 0.28265448 0.25726578
#> 97       0.44 0.86488543 0.84830582
#> 98       0.44 0.53585499 0.53279774
#> 99       0.44 0.25289149 0.23085503
#> 100      0.44 0.64723712 0.61743261