Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
#> Multiple minima.
predict(net,newx=X)
#>     intercept  gaussian  binomial  combined    switch
#> 1        0.54 0.7047312 0.6453970 0.6844033 0.6453970
#> 2        0.54 0.6078730 0.7539811 0.7808711 0.7539811
#> 3        0.54 0.6995445 0.6868910 0.7213330 0.6868910
#> 4        0.54 0.2259400 0.1380540 0.1228680 0.1380540
#> 5        0.54 0.2807055 0.3331342 0.2964895 0.3331342
#> 6        0.54 0.4300282 0.4886921 0.4380753 0.4886921
#> 7        0.54 0.8144714 0.8630044 0.8780739 0.8630044
#> 8        0.54 0.3030782 0.1843755 0.1640942 0.1843755
#> 9        0.54 0.8377272 0.8481702 0.8648714 0.8481702
#> 10       0.54 0.6007557 0.8082695 0.8290374 0.8082695
#> 11       0.54 0.8080131 0.9497184 0.9552494 0.9497184
#> 12       0.54 0.4177896 0.2348846 0.2104274 0.2348846
#> 13       0.54 0.2142374 0.1659555 0.1477004 0.1659555
#> 14       0.54 0.4129709 0.1835299 0.1643112 0.1835299
#> 15       0.54 0.7678711 0.7334976 0.7628128 0.7334976
#> 16       0.54 0.2034168 0.1569548 0.1396897 0.1569548
#> 17       0.54 0.4155450 0.3320948 0.2967376 0.3320948
#> 18       0.54 0.7089543 0.8528436 0.8690308 0.8528436
#> 19       0.54 0.6901085 0.3836520 0.4514503 0.3836520
#> 20       0.54 0.7637795 0.8923063 0.9041526 0.8923063
#> 21       0.54 0.7302340 0.6371245 0.6770408 0.6371245
#> 22       0.54 0.5647640 0.9002310 0.9068912 0.9002310
#> 23       0.54 0.1651122 0.2199591 0.1957636 0.2199591
#> 24       0.54 0.2251203 0.3834826 0.3412995 0.3834826
#> 25       0.54 0.6148214 0.8299140 0.8485338 0.8299140
#> 26       0.54 0.2417794 0.2392467 0.2129296 0.2392467
#> 27       0.54 0.4464022 0.5607593 0.5070977 0.5607593
#> 28       0.54 0.4617604 0.5476891 0.5039528 0.5476891
#> 29       0.54 0.6691118 0.6609161 0.6982152 0.6609161
#> 30       0.54 0.7519411 0.8783904 0.8917674 0.8783904
#> 31       0.54 0.5010008 0.7647755 0.7368381 0.7647755
#> 32       0.54 0.7137305 0.5454121 0.5954168 0.5454121
#> 33       0.54 0.4884447 0.5372338 0.5196385 0.5372338
#> 34       0.54 0.5316146 0.6174377 0.6379642 0.6174377
#> 35       0.54 0.3224941 0.5009910 0.4458820 0.5009910
#> 36       0.54 0.6586922 0.8107013 0.8315236 0.8107013
#> 37       0.54 0.4002183 0.5406456 0.4815249 0.5406456
#> 38       0.54 0.6235016 0.6309703 0.6715260 0.6309703
#> 39       0.54 0.6926355 0.7242382 0.7545720 0.7242382
#> 40       0.54 0.4355124 0.6750181 0.6051517 0.6750181
#> 41       0.54 0.7577134 0.8472177 0.8640237 0.8472177
#> 42       0.54 0.3269590 0.3155764 0.2808631 0.3155764
#> 43       0.54 0.5856861 0.4764620 0.5329796 0.4764620
#> 44       0.54 0.4481855 0.4143460 0.3775582 0.4143460
#> 45       0.54 0.2990999 0.3467162 0.3085775 0.3467162
#> 46       0.54 0.4311540 0.7465648 0.6678107 0.7465648
#> 47       0.54 0.6759184 0.6470183 0.6858463 0.6470183
#> 48       0.54 0.2628453 0.2671057 0.2377241 0.2671057
#> 49       0.54 0.6652885 0.5944573 0.6390668 0.5944573
#> 50       0.54 0.4975232 0.5235680 0.5180373 0.5235680
#> 51       0.54 0.7862855 0.7441312 0.7722767 0.7441312
#> 52       0.54 0.4712188 0.4799575 0.4511321 0.4799575
#> 53       0.54 0.5065641 0.4125605 0.4299311 0.4125605
#> 54       0.54 0.4970098 0.6513802 0.6311824 0.6513802
#> 55       0.54 0.6926125 0.7085389 0.7405996 0.7085389
#> 56       0.54 0.5767109 0.7183057 0.7472690 0.7183057
#> 57       0.54 0.2172549 0.3534159 0.3145401 0.3534159
#> 58       0.54 0.1901227 0.1505254 0.1339676 0.1505254
#> 59       0.54 0.6202759 0.8160882 0.8362662 0.8160882
#> 60       0.54 0.5471465 0.6894238 0.7125264 0.6894238
#> 61       0.54 0.7605441 0.5057538 0.5601209 0.5057538
#> 62       0.54 0.3131477 0.2102984 0.1871656 0.2102984
#> 63       0.54 0.6093500 0.4844983 0.5410532 0.4844983
#> 64       0.54 0.4115940 0.3076009 0.2746386 0.3076009
#> 65       0.54 0.5113986 0.3535792 0.3830078 0.3535792
#> 66       0.54 0.1568087 0.2652558 0.2360777 0.2652558
#> 67       0.54 0.4013218 0.6974229 0.6210909 0.6974229
#> 68       0.54 0.5761531 0.7240366 0.7522921 0.7240366
#> 69       0.54 0.5137445 0.4417224 0.4640846 0.4417224
#> 70       0.54 0.7938699 0.8332653 0.8516061 0.8332653
#> 71       0.54 0.5890419 0.8289993 0.8469770 0.8289993
#> 72       0.54 0.5745155 0.5245917 0.5745440 0.5245917
#> 73       0.54 0.4718017 0.5516435 0.5154479 0.5516435
#> 74       0.54 0.5613792 0.4477250 0.5032226 0.4477250
#> 75       0.54 0.3425745 0.3525055 0.3137306 0.3525055
#> 76       0.54 0.6865887 0.4428951 0.5041766 0.4428951
#> 77       0.54 0.5347742 0.5028494 0.5384931 0.5028494
#> 78       0.54 0.4440747 0.2738487 0.2508225 0.2738487
#> 79       0.54 0.6342415 0.7395892 0.7682226 0.7395892
#> 80       0.54 0.5876552 0.6770710 0.7116682 0.6770710
#> 81       0.54 0.7079223 0.7569145 0.7836539 0.7569145
#> 82       0.54 0.3344354 0.2035401 0.1811509 0.2035401
#> 83       0.54 0.6320484 0.4769419 0.5344633 0.4769419
#> 84       0.54 0.2474321 0.2740969 0.2439463 0.2740969
#> 85       0.54 0.6230524 0.7328408 0.7621889 0.7328408
#> 86       0.54 0.6894927 0.4631751 0.5222258 0.4631751
#> 87       0.54 0.7821074 0.7015335 0.7343648 0.7015335
#> 88       0.54 0.3976546 0.3251203 0.2896384 0.3251203
#> 89       0.54 0.5045699 0.5781266 0.5749443 0.5781266
#> 90       0.54 0.3283070 0.5203143 0.4630799 0.5203143
#> 91       0.54 0.4779070 0.6560849 0.6141576 0.6560849
#> 92       0.54 0.8131419 0.7622504 0.7884028 0.7622504
#> 93       0.54 0.5938642 0.6247155 0.6654263 0.6247155
#> 94       0.54 0.8483995 0.8560982 0.8719274 0.8560982
#> 95       0.54 0.4916966 0.2819775 0.2961840 0.2819775
#> 96       0.54 0.2054682 0.1685791 0.1500354 0.1685791
#> 97       0.54 0.4962698 0.4148962 0.4198367 0.4148962
#> 98       0.54 0.3893705 0.6352259 0.5654846 0.6352259
#> 99       0.54 0.4064333 0.3519159 0.3137894 0.3519159
#> 100      0.54 0.5667677 0.3065977 0.3790460 0.3065977