Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for class 'cornet'
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
#> Multiple minima.
predict(net,newx=X)
#>     intercept  gaussian  binomial  combined    switch
#> 1        0.48 0.4313983 0.3844532 0.3472188 0.3844532
#> 2        0.48 0.4247109 0.4644889 0.4108327 0.4644889
#> 3        0.48 0.4321706 0.4198350 0.3773819 0.4198350
#> 4        0.48 0.4234110 0.5087978 0.4473939 0.5087978
#> 5        0.48 0.4250316 0.4649304 0.4113682 0.4649304
#> 6        0.48 0.4248726 0.2585689 0.2379428 0.2585689
#> 7        0.48 0.4319027 0.3943015 0.3557797 0.3943015
#> 8        0.48 0.4264054 0.2976522 0.2715698 0.2976522
#> 9        0.48 0.4356959 0.5516685 0.4902068 0.5516685
#> 10       0.48 0.4297888 0.5200674 0.4602292 0.5200674
#> 11       0.48 0.4294778 0.4767300 0.4236534 0.4767300
#> 12       0.48 0.4214309 0.5028575 0.4414288 0.5028575
#> 13       0.48 0.4143182 0.3932802 0.3461519 0.3932802
#> 14       0.48 0.4261999 0.2826256 0.2588394 0.2826256
#> 15       0.48 0.4237447 0.5420910 0.4755278 0.5420910
#> 16       0.48 0.4445638 0.6160975 0.5500410 0.6160975
#> 17       0.48 0.4210046 0.4885728 0.4292240 0.4885728
#> 18       0.48 0.4336210 0.4218429 0.3799133 0.4218429
#> 19       0.48 0.4289162 0.3487328 0.3158266 0.3487328
#> 20       0.48 0.4503444 0.5511590 0.4995694 0.5511590
#> 21       0.48 0.4370105 0.3986516 0.3624774 0.3986516
#> 22       0.48 0.4313684 0.6232158 0.5477623 0.6232158
#> 23       0.48 0.4244529 0.5642535 0.4945032 0.5642535
#> 24       0.48 0.4318151 0.5932565 0.5228516 0.5932565
#> 25       0.48 0.4223419 0.3592862 0.3212735 0.3592862
#> 26       0.48 0.4419264 0.3898746 0.3582450 0.3898746
#> 27       0.48 0.4325921 0.5963688 0.5259139 0.5963688
#> 28       0.48 0.4255029 0.4020030 0.3587527 0.4020030
#> 29       0.48 0.4291088 0.3905839 0.3510873 0.3905839
#> 30       0.48 0.4324075 0.4231594 0.3803111 0.4231594
#> 31       0.48 0.4399888 0.3563697 0.3288386 0.3563697
#> 32       0.48 0.4332571 0.6887659 0.6039150 0.6887659
#> 33       0.48 0.4233281 0.4246893 0.3767012 0.4246893
#> 34       0.48 0.4178994 0.5141670 0.4492715 0.5141670
#> 35       0.48 0.4193387 0.4261104 0.3759668 0.4261104
#> 36       0.48 0.4285248 0.4858044 0.4307529 0.4858044
#> 37       0.48 0.4225714 0.5298829 0.4646878 0.5298829
#> 38       0.48 0.4272446 0.5055782 0.4466725 0.5055782
#> 39       0.48 0.4296154 0.5953436 0.5233650 0.5953436
#> 40       0.48 0.4396389 0.4899631 0.4408325 0.4899631
#> 41       0.48 0.4161463 0.3901367 0.3443021 0.3901367
#> 42       0.48 0.4332394 0.5415188 0.4802170 0.5415188
#> 43       0.48 0.4369775 0.6081723 0.5384544 0.6081723
#> 44       0.48 0.4267367 0.5347677 0.4709217 0.5347677
#> 45       0.48 0.4150528 0.2592411 0.2338735 0.2592411
#> 46       0.48 0.4269127 0.4234990 0.3775493 0.4234990
#> 47       0.48 0.4222714 0.5654822 0.4944434 0.5654822
#> 48       0.48 0.4220849 0.6448528 0.5610232 0.6448528
#> 49       0.48 0.4366654 0.6578827 0.5800188 0.6578827
#> 50       0.48 0.4171396 0.3564770 0.3164690 0.3564770
#> 51       0.48 0.4236629 0.3861827 0.3445237 0.3861827
#> 52       0.48 0.4337232 0.4603841 0.4123481 0.4603841
#> 53       0.48 0.4146946 0.3541962 0.3134819 0.3541962
#> 54       0.48 0.4168985 0.3619044 0.3209202 0.3619044
#> 55       0.48 0.4222182 0.4843761 0.4262882 0.4843761
#> 56       0.48 0.4320380 0.4455148 0.3988765 0.4455148
#> 57       0.48 0.4422194 0.3623156 0.3352891 0.3623156
#> 58       0.48 0.4272073 0.4508363 0.4006694 0.4508363
#> 59       0.48 0.4253257 0.4128668 0.3677865 0.4128668
#> 60       0.48 0.4253932 0.5342947 0.4698209 0.5342947
#> 61       0.48 0.4264306 0.3950293 0.3533798 0.3950293
#> 62       0.48 0.4246221 0.4178507 0.3716112 0.4178507
#> 63       0.48 0.4158342 0.4788605 0.4186931 0.4788605
#> 64       0.48 0.4254526 0.4187001 0.3727522 0.4187001
#> 65       0.48 0.4241666 0.5182120 0.4556827 0.5182120
#> 66       0.48 0.4162062 0.5226454 0.4556357 0.5226454
#> 67       0.48 0.4291737 0.5210492 0.4607137 0.5210492
#> 68       0.48 0.4329938 0.5755354 0.5086475 0.5755354
#> 69       0.48 0.4210918 0.4446821 0.3923978 0.4446821
#> 70       0.48 0.4325649 0.4569095 0.4087523 0.4569095
#> 71       0.48 0.4212734 0.4545708 0.4007918 0.4545708
#> 72       0.48 0.4273380 0.5854647 0.5138271 0.5854647
#> 73       0.48 0.4212236 0.4790256 0.4213099 0.4790256
#> 74       0.48 0.4303846 0.4970932 0.4412635 0.4970932
#> 75       0.48 0.4333243 0.6291282 0.5538587 0.6291282
#> 76       0.48 0.4103131 0.5405638 0.4682324 0.5405638
#> 77       0.48 0.4352502 0.6687256 0.5882653 0.6687256
#> 78       0.48 0.4067512 0.3285039 0.2887521 0.3285039
#> 79       0.48 0.4341310 0.5473756 0.4856621 0.5473756
#> 80       0.48 0.4339529 0.5185549 0.4613472 0.5185549
#> 81       0.48 0.4428646 0.6089664 0.5429046 0.6089664
#> 82       0.48 0.4287063 0.3136868 0.2862731 0.3136868
#> 83       0.48 0.4345031 0.4668503 0.4182422 0.4668503
#> 84       0.48 0.4316001 0.4641955 0.4143174 0.4641955
#> 85       0.48 0.4263722 0.4584878 0.4066542 0.4584878
#> 86       0.48 0.4265849 0.4737707 0.4196039 0.4737707
#> 87       0.48 0.4227184 0.4735713 0.4174587 0.4735713
#> 88       0.48 0.4193210 0.3452349 0.3080232 0.3452349
#> 89       0.48 0.4283075 0.5582076 0.4914534 0.5582076
#> 90       0.48 0.4291490 0.4033744 0.3618533 0.4033744
#> 91       0.48 0.4310192 0.5102519 0.4526744 0.5102519
#> 92       0.48 0.4210150 0.4410029 0.3892703 0.4410029
#> 93       0.48 0.4408505 0.6323865 0.5612502 0.6323865
#> 94       0.48 0.4383190 0.5652644 0.5032480 0.5652644
#> 95       0.48 0.4248393 0.3963812 0.3536880 0.3963812
#> 96       0.48 0.4249580 0.6999208 0.6087223 0.6999208
#> 97       0.48 0.4318486 0.5503005 0.4867878 0.5503005
#> 98       0.48 0.4162194 0.4187137 0.3683389 0.4187137
#> 99       0.48 0.4347387 0.6556977 0.5770147 0.6556977
#> 100      0.48 0.4340159 0.6842714 0.6005864 0.6842714