Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)
#>     intercept binomial  combined
#> 1         0.5      0.5 0.5008209
#> 2         0.5      0.5 0.5008209
#> 3         0.5      0.5 0.5008209
#> 4         0.5      0.5 0.5008209
#> 5         0.5      0.5 0.5008209
#> 6         0.5      0.5 0.5008209
#> 7         0.5      0.5 0.5008209
#> 8         0.5      0.5 0.5008209
#> 9         0.5      0.5 0.5008209
#> 10        0.5      0.5 0.5008209
#> 11        0.5      0.5 0.5008209
#> 12        0.5      0.5 0.5008209
#> 13        0.5      0.5 0.5008209
#> 14        0.5      0.5 0.5008209
#> 15        0.5      0.5 0.5008209
#> 16        0.5      0.5 0.5008209
#> 17        0.5      0.5 0.5008209
#> 18        0.5      0.5 0.5008209
#> 19        0.5      0.5 0.5008209
#> 20        0.5      0.5 0.5008209
#> 21        0.5      0.5 0.5008209
#> 22        0.5      0.5 0.5008209
#> 23        0.5      0.5 0.5008209
#> 24        0.5      0.5 0.5008209
#> 25        0.5      0.5 0.5008209
#> 26        0.5      0.5 0.5008209
#> 27        0.5      0.5 0.5008209
#> 28        0.5      0.5 0.5008209
#> 29        0.5      0.5 0.5008209
#> 30        0.5      0.5 0.5008209
#> 31        0.5      0.5 0.5008209
#> 32        0.5      0.5 0.5008209
#> 33        0.5      0.5 0.5008209
#> 34        0.5      0.5 0.5008209
#> 35        0.5      0.5 0.5008209
#> 36        0.5      0.5 0.5008209
#> 37        0.5      0.5 0.5008209
#> 38        0.5      0.5 0.5008209
#> 39        0.5      0.5 0.5008209
#> 40        0.5      0.5 0.5008209
#> 41        0.5      0.5 0.5008209
#> 42        0.5      0.5 0.5008209
#> 43        0.5      0.5 0.5008209
#> 44        0.5      0.5 0.5008209
#> 45        0.5      0.5 0.5008209
#> 46        0.5      0.5 0.5008209
#> 47        0.5      0.5 0.5008209
#> 48        0.5      0.5 0.5008209
#> 49        0.5      0.5 0.5008209
#> 50        0.5      0.5 0.5008209
#> 51        0.5      0.5 0.5008209
#> 52        0.5      0.5 0.5008209
#> 53        0.5      0.5 0.5008209
#> 54        0.5      0.5 0.5008209
#> 55        0.5      0.5 0.5008209
#> 56        0.5      0.5 0.5008209
#> 57        0.5      0.5 0.5008209
#> 58        0.5      0.5 0.5008209
#> 59        0.5      0.5 0.5008209
#> 60        0.5      0.5 0.5008209
#> 61        0.5      0.5 0.5008209
#> 62        0.5      0.5 0.5008209
#> 63        0.5      0.5 0.5008209
#> 64        0.5      0.5 0.5008209
#> 65        0.5      0.5 0.5008209
#> 66        0.5      0.5 0.5008209
#> 67        0.5      0.5 0.5008209
#> 68        0.5      0.5 0.5008209
#> 69        0.5      0.5 0.5008209
#> 70        0.5      0.5 0.5008209
#> 71        0.5      0.5 0.5008209
#> 72        0.5      0.5 0.5008209
#> 73        0.5      0.5 0.5008209
#> 74        0.5      0.5 0.5008209
#> 75        0.5      0.5 0.5008209
#> 76        0.5      0.5 0.5008209
#> 77        0.5      0.5 0.5008209
#> 78        0.5      0.5 0.5008209
#> 79        0.5      0.5 0.5008209
#> 80        0.5      0.5 0.5008209
#> 81        0.5      0.5 0.5008209
#> 82        0.5      0.5 0.5008209
#> 83        0.5      0.5 0.5008209
#> 84        0.5      0.5 0.5008209
#> 85        0.5      0.5 0.5008209
#> 86        0.5      0.5 0.5008209
#> 87        0.5      0.5 0.5008209
#> 88        0.5      0.5 0.5008209
#> 89        0.5      0.5 0.5008209
#> 90        0.5      0.5 0.5008209
#> 91        0.5      0.5 0.5008209
#> 92        0.5      0.5 0.5008209
#> 93        0.5      0.5 0.5008209
#> 94        0.5      0.5 0.5008209
#> 95        0.5      0.5 0.5008209
#> 96        0.5      0.5 0.5008209
#> 97        0.5      0.5 0.5008209
#> 98        0.5      0.5 0.5008209
#> 99        0.5      0.5 0.5008209
#> 100       0.5      0.5 0.5008209