Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)
#>     intercept  gaussian binomial  combined    switch
#> 1        0.46 0.5296598     0.46 0.6023390 0.5296598
#> 2        0.46 0.4338613     0.46 0.3012592 0.4338613
#> 3        0.46 0.4153503     0.46 0.2992476 0.4153503
#> 4        0.46 0.5166983     0.46 0.5559058 0.5166983
#> 5        0.46 0.5709174     0.46 0.6476637 0.5709174
#> 6        0.46 0.5820455     0.46 0.6486520 0.5820455
#> 7        0.46 0.3848275     0.46 0.2990022 0.3848275
#> 8        0.46 0.4807938     0.46 0.3816308 0.4807938
#> 9        0.46 0.5093272     0.46 0.5217646 0.5093272
#> 10       0.46 0.4661277     0.46 0.3347788 0.4661277
#> 11       0.46 0.4868303     0.46 0.4078672 0.4868303
#> 12       0.46 0.5222144     0.46 0.5780196 0.5222144
#> 13       0.46 0.4131848     0.46 0.2991852 0.4131848
#> 14       0.46 0.4925627     0.46 0.4356358 0.4925627
#> 15       0.46 0.5131161     0.46 0.5398844 0.5131161
#> 16       0.46 0.4576317     0.46 0.3186365 0.4576317
#> 17       0.46 0.5006118     0.46 0.4771960 0.5006118
#> 18       0.46 0.5569891     0.46 0.6433343 0.5569891
#> 19       0.46 0.4772948     0.46 0.3681768 0.4772948
#> 20       0.46 0.5908826     0.46 0.6488946 0.5908826
#> 21       0.46 0.4240467     0.46 0.2997427 0.4240467
#> 22       0.46 0.3687738     0.46 0.2990001 0.3687738
#> 23       0.46 0.5435084     0.46 0.6310036 0.5435084
#> 24       0.46 0.3869595     0.46 0.2990032 0.3869595
#> 25       0.46 0.5376670     0.46 0.6213349 0.5376670
#> 26       0.46 0.4089879     0.46 0.2991035 0.4089879
#> 27       0.46 0.4869090     0.46 0.4082320 0.4869090
#> 28       0.46 0.5378655     0.46 0.6217175 0.5378655
#> 29       0.46 0.3584875     0.46 0.2990000 0.3584875
#> 30       0.46 0.4818492     0.46 0.3859539 0.4818492
#> 31       0.46 0.4795855     0.46 0.3768299 0.4795855
#> 32       0.46 0.4355270     0.46 0.3016932 0.4355270
#> 33       0.46 0.5149846     0.46 0.5483903 0.5149846
#> 34       0.46 0.4302703     0.46 0.3005271 0.4302703
#> 35       0.46 0.5275877     0.46 0.5962284 0.5275877
#> 36       0.46 0.4210495     0.46 0.2995148 0.4210495
#> 37       0.46 0.4267404     0.46 0.3000213 0.4267404
#> 38       0.46 0.4750381     0.46 0.3602452 0.4750381
#> 39       0.46 0.4881313     0.46 0.4139608 0.4881313
#> 40       0.46 0.5565857     0.46 0.6431150 0.5565857
#> 41       0.46 0.5429188     0.46 0.6301696 0.5429188
#> 42       0.46 0.4315846     0.46 0.3007661 0.4315846
#> 43       0.46 0.4689617     0.46 0.3418752 0.4689617
#> 44       0.46 0.4838817     0.46 0.3946059 0.4838817
#> 45       0.46 0.4655082     0.46 0.3333480 0.4655082
#> 46       0.46 0.4176474     0.46 0.2993344 0.4176474
#> 47       0.46 0.5441624     0.46 0.6318943 0.5441624
#> 48       0.46 0.3924547     0.46 0.2990081 0.3924547
#> 49       0.46 0.4927926     0.46 0.4367928 0.4927926
#> 50       0.46 0.3903830     0.46 0.2990057 0.3903830
#> 51       0.46 0.3768332     0.46 0.2990005 0.3768332
#> 52       0.46 0.4299248     0.46 0.3004692 0.4299248
#> 53       0.46 0.5196028     0.46 0.5679626 0.5196028
#> 54       0.46 0.4641585     0.46 0.3303760 0.4641585
#> 55       0.46 0.4844339     0.46 0.3970279 0.4844339
#> 56       0.46 0.5978762     0.46 0.6489623 0.5978762
#> 57       0.46 0.5138780     0.46 0.5433905 0.5138780
#> 58       0.46 0.4789349     0.46 0.3743116 0.4789349
#> 59       0.46 0.5596776     0.46 0.6446253 0.5596776
#> 60       0.46 0.4596288     0.46 0.3217896 0.4596288
#> 61       0.46 0.5888219     0.46 0.6488594 0.5888219
#> 62       0.46 0.4756286     0.46 0.3622630 0.4756286
#> 63       0.46 0.3328821     0.46 0.2990000 0.3328821
#> 64       0.46 0.4516413     0.46 0.3111947 0.4516413
#> 65       0.46 0.5171598     0.46 0.5578800 0.5171598
#> 66       0.46 0.5677125     0.46 0.6470930 0.5677125
#> 67       0.46 0.4612974     0.46 0.3247124 0.4612974
#> 68       0.46 0.4067655     0.46 0.2990752 0.4067655
#> 69       0.46 0.5134469     0.46 0.5414127 0.5134469
#> 70       0.46 0.5273859     0.46 0.5956063 0.5273859
#> 71       0.46 0.5071654     0.46 0.5109952 0.5071654
#> 72       0.46 0.4969728     0.46 0.4582171 0.4969728
#> 73       0.46 0.5073662     0.46 0.5120064 0.5073662
#> 74       0.46 0.4608997     0.46 0.3239911 0.4608997
#> 75       0.46 0.3207417     0.46 0.2990000 0.3207417
#> 76       0.46 0.4145551     0.46 0.2992227 0.4145551
#> 77       0.46 0.4939152     0.46 0.4424800 0.4939152
#> 78       0.46 0.3930854     0.46 0.2990090 0.3930854
#> 79       0.46 0.6054711     0.46 0.6489886 0.6054711
#> 80       0.46 0.4222950     0.46 0.2996004 0.4222950
#> 81       0.46 0.4689028     0.46 0.3417183 0.4689028
#> 82       0.46 0.4519239     0.46 0.3114845 0.4519239
#> 83       0.46 0.4464886     0.46 0.3068060 0.4464886
#> 84       0.46 0.5208178     0.46 0.5727368 0.5208178
#> 85       0.46 0.3821992     0.46 0.2990013 0.3821992
#> 86       0.46 0.4684317     0.46 0.3404785 0.4684317
#> 87       0.46 0.3366680     0.46 0.2990000 0.3366680
#> 88       0.46 0.4722639     0.46 0.3513172 0.4722639
#> 89       0.46 0.4919012     0.46 0.4323219 0.4919012
#> 90       0.46 0.3860032     0.46 0.2990027 0.3860032
#> 91       0.46 0.5217779     0.46 0.5763922 0.5217779
#> 92       0.46 0.5027173     0.46 0.4881729 0.5027173
#> 93       0.46 0.5056732     0.46 0.5034209 0.5056732
#> 94       0.46 0.4681578     0.46 0.3397693 0.4681578
#> 95       0.46 0.5399159     0.46 0.6254396 0.5399159
#> 96       0.46 0.4932190     0.46 0.4389460 0.4932190
#> 97       0.46 0.4656158     0.46 0.3335935 0.4656158
#> 98       0.46 0.6280579     0.46 0.6489998 0.6280579
#> 99       0.46 0.5268189     0.46 0.5938320 0.5268189
#> 100      0.46 0.5029504     0.46 0.4893839 0.5029504