Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object
cornet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
"probability", "odds", "log-odds"

...
further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)

#>     intercept binomial  combined
#> 1        0.48     0.48 0.4454032
#> 2        0.48     0.48 0.7199833
#> 3        0.48     0.48 0.5093483
#> 4        0.48     0.48 0.4908294
#> 5        0.48     0.48 0.4388327
#> 6        0.48     0.48 0.4412961
#> 7        0.48     0.48 0.2428037
#> 8        0.48     0.48 0.2072323
#> 9        0.48     0.48 0.4558324
#> 10       0.48     0.48 0.3392687
#> 11       0.48     0.48 0.6512743
#> 12       0.48     0.48 0.4874262
#> 13       0.48     0.48 0.5182075
#> 14       0.48     0.48 0.4112037
#> 15       0.48     0.48 0.4127740
#> 16       0.48     0.48 0.3526974
#> 17       0.48     0.48 0.5921134
#> 18       0.48     0.48 0.5833832
#> 19       0.48     0.48 0.5416940
#> 20       0.48     0.48 0.4019544
#> 21       0.48     0.48 0.5365248
#> 22       0.48     0.48 0.6827133
#> 23       0.48     0.48 0.3892998
#> 24       0.48     0.48 0.5702544
#> 25       0.48     0.48 0.5127992
#> 26       0.48     0.48 0.5066469
#> 27       0.48     0.48 0.4829527
#> 28       0.48     0.48 0.4661661
#> 29       0.48     0.48 0.4417929
#> 30       0.48     0.48 0.4640288
#> 31       0.48     0.48 0.5749693
#> 32       0.48     0.48 0.4687028
#> 33       0.48     0.48 0.5150003
#> 34       0.48     0.48 0.4874637
#> 35       0.48     0.48 0.6466244
#> 36       0.48     0.48 0.4100652
#> 37       0.48     0.48 0.5427689
#> 38       0.48     0.48 0.6039042
#> 39       0.48     0.48 0.6353845
#> 40       0.48     0.48 0.4836715
#> 41       0.48     0.48 0.5810545
#> 42       0.48     0.48 0.4690693
#> 43       0.48     0.48 0.3016047
#> 44       0.48     0.48 0.4938419
#> 45       0.48     0.48 0.2757423
#> 46       0.48     0.48 0.5238304
#> 47       0.48     0.48 0.6084295
#> 48       0.48     0.48 0.3361513
#> 49       0.48     0.48 0.5209240
#> 50       0.48     0.48 0.4708944
#> 51       0.48     0.48 0.4004816
#> 52       0.48     0.48 0.3741752
#> 53       0.48     0.48 0.2559125
#> 54       0.48     0.48 0.4944509
#> 55       0.48     0.48 0.4401203
#> 56       0.48     0.48 0.4696028
#> 57       0.48     0.48 0.5088430
#> 58       0.48     0.48 0.5071365
#> 59       0.48     0.48 0.4372763
#> 60       0.48     0.48 0.3706754
#> 61       0.48     0.48 0.5344520
#> 62       0.48     0.48 0.6375873
#> 63       0.48     0.48 0.4686650
#> 64       0.48     0.48 0.5335537
#> 65       0.48     0.48 0.4044528
#> 66       0.48     0.48 0.2338702
#> 67       0.48     0.48 0.3969799
#> 68       0.48     0.48 0.6999328
#> 69       0.48     0.48 0.3183644
#> 70       0.48     0.48 0.4140908
#> 71       0.48     0.48 0.3885233
#> 72       0.48     0.48 0.4753442
#> 73       0.48     0.48 0.4266898
#> 74       0.48     0.48 0.4195858
#> 75       0.48     0.48 0.4935820
#> 76       0.48     0.48 0.3171190
#> 77       0.48     0.48 0.6599498
#> 78       0.48     0.48 0.2725207
#> 79       0.48     0.48 0.3627156
#> 80       0.48     0.48 0.6589043
#> 81       0.48     0.48 0.5511269
#> 82       0.48     0.48 0.3552917
#> 83       0.48     0.48 0.2824604
#> 84       0.48     0.48 0.2434102
#> 85       0.48     0.48 0.2461497
#> 86       0.48     0.48 0.5233560
#> 87       0.48     0.48 0.4498209
#> 88       0.48     0.48 0.5126292
#> 89       0.48     0.48 0.6459748
#> 90       0.48     0.48 0.3486921
#> 91       0.48     0.48 0.4659105
#> 92       0.48     0.48 0.3819994
#> 93       0.48     0.48 0.4878090
#> 94       0.48     0.48 0.4968532
#> 95       0.48     0.48 0.3961838
#> 96       0.48     0.48 0.4922362
#> 97       0.48     0.48 0.4075505
#> 98       0.48     0.48 0.6393087
#> 99       0.48     0.48 0.3336027
#> 100      0.48     0.48 0.5511489