Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object
cornet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
"probability", "odds", "log-odds"

...
further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)

#>     intercept binomial  combined
#> 1        0.45     0.45 0.4486933
#> 2        0.45     0.45 0.4486933
#> 3        0.45     0.45 0.4486933
#> 4        0.45     0.45 0.4486933
#> 5        0.45     0.45 0.4486933
#> 6        0.45     0.45 0.4486933
#> 7        0.45     0.45 0.4486933
#> 8        0.45     0.45 0.4486933
#> 9        0.45     0.45 0.4486933
#> 10       0.45     0.45 0.4486933
#> 11       0.45     0.45 0.4486933
#> 12       0.45     0.45 0.4486933
#> 13       0.45     0.45 0.4486933
#> 14       0.45     0.45 0.4486933
#> 15       0.45     0.45 0.4486933
#> 16       0.45     0.45 0.4486933
#> 17       0.45     0.45 0.4486933
#> 18       0.45     0.45 0.4486933
#> 19       0.45     0.45 0.4486933
#> 20       0.45     0.45 0.4486933
#> 21       0.45     0.45 0.4486933
#> 22       0.45     0.45 0.4486933
#> 23       0.45     0.45 0.4486933
#> 24       0.45     0.45 0.4486933
#> 25       0.45     0.45 0.4486933
#> 26       0.45     0.45 0.4486933
#> 27       0.45     0.45 0.4486933
#> 28       0.45     0.45 0.4486933
#> 29       0.45     0.45 0.4486933
#> 30       0.45     0.45 0.4486933
#> 31       0.45     0.45 0.4486933
#> 32       0.45     0.45 0.4486933
#> 33       0.45     0.45 0.4486933
#> 34       0.45     0.45 0.4486933
#> 35       0.45     0.45 0.4486933
#> 36       0.45     0.45 0.4486933
#> 37       0.45     0.45 0.4486933
#> 38       0.45     0.45 0.4486933
#> 39       0.45     0.45 0.4486933
#> 40       0.45     0.45 0.4486933
#> 41       0.45     0.45 0.4486933
#> 42       0.45     0.45 0.4486933
#> 43       0.45     0.45 0.4486933
#> 44       0.45     0.45 0.4486933
#> 45       0.45     0.45 0.4486933
#> 46       0.45     0.45 0.4486933
#> 47       0.45     0.45 0.4486933
#> 48       0.45     0.45 0.4486933
#> 49       0.45     0.45 0.4486933
#> 50       0.45     0.45 0.4486933
#> 51       0.45     0.45 0.4486933
#> 52       0.45     0.45 0.4486933
#> 53       0.45     0.45 0.4486933
#> 54       0.45     0.45 0.4486933
#> 55       0.45     0.45 0.4486933
#> 56       0.45     0.45 0.4486933
#> 57       0.45     0.45 0.4486933
#> 58       0.45     0.45 0.4486933
#> 59       0.45     0.45 0.4486933
#> 60       0.45     0.45 0.4486933
#> 61       0.45     0.45 0.4486933
#> 62       0.45     0.45 0.4486933
#> 63       0.45     0.45 0.4486933
#> 64       0.45     0.45 0.4486933
#> 65       0.45     0.45 0.4486933
#> 66       0.45     0.45 0.4486933
#> 67       0.45     0.45 0.4486933
#> 68       0.45     0.45 0.4486933
#> 69       0.45     0.45 0.4486933
#> 70       0.45     0.45 0.4486933
#> 71       0.45     0.45 0.4486933
#> 72       0.45     0.45 0.4486933
#> 73       0.45     0.45 0.4486933
#> 74       0.45     0.45 0.4486933
#> 75       0.45     0.45 0.4486933
#> 76       0.45     0.45 0.4486933
#> 77       0.45     0.45 0.4486933
#> 78       0.45     0.45 0.4486933
#> 79       0.45     0.45 0.4486933
#> 80       0.45     0.45 0.4486933
#> 81       0.45     0.45 0.4486933
#> 82       0.45     0.45 0.4486933
#> 83       0.45     0.45 0.4486933
#> 84       0.45     0.45 0.4486933
#> 85       0.45     0.45 0.4486933
#> 86       0.45     0.45 0.4486933
#> 87       0.45     0.45 0.4486933
#> 88       0.45     0.45 0.4486933
#> 89       0.45     0.45 0.4486933
#> 90       0.45     0.45 0.4486933
#> 91       0.45     0.45 0.4486933
#> 92       0.45     0.45 0.4486933
#> 93       0.45     0.45 0.4486933
#> 94       0.45     0.45 0.4486933
#> 95       0.45     0.45 0.4486933
#> 96       0.45     0.45 0.4486933
#> 97       0.45     0.45 0.4486933
#> 98       0.45     0.45 0.4486933
#> 99       0.45     0.45 0.4486933
#> 100      0.45     0.45 0.4486933