Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)
#>     intercept binomial  combined
#> 1        0.51     0.51 0.4262646
#> 2        0.51     0.51 0.6143782
#> 3        0.51     0.51 0.3983528
#> 4        0.51     0.51 0.6883158
#> 5        0.51     0.51 0.4925129
#> 6        0.51     0.51 0.3430459
#> 7        0.51     0.51 0.7030838
#> 8        0.51     0.51 0.6711004
#> 9        0.51     0.51 0.4566793
#> 10       0.51     0.51 0.5427023
#> 11       0.51     0.51 0.5667485
#> 12       0.51     0.51 0.3233813
#> 13       0.51     0.51 0.6155116
#> 14       0.51     0.51 0.5877398
#> 15       0.51     0.51 0.4809649
#> 16       0.51     0.51 0.4605397
#> 17       0.51     0.51 0.5902571
#> 18       0.51     0.51 0.7365414
#> 19       0.51     0.51 0.3487755
#> 20       0.51     0.51 0.7344896
#> 21       0.51     0.51 0.5776025
#> 22       0.51     0.51 0.3597236
#> 23       0.51     0.51 0.6110098
#> 24       0.51     0.51 0.6490268
#> 25       0.51     0.51 0.7061371
#> 26       0.51     0.51 0.4679033
#> 27       0.51     0.51 0.4969595
#> 28       0.51     0.51 0.4786304
#> 29       0.51     0.51 0.5852548
#> 30       0.51     0.51 0.3110540
#> 31       0.51     0.51 0.5179571
#> 32       0.51     0.51 0.4891834
#> 33       0.51     0.51 0.1738792
#> 34       0.51     0.51 0.6204240
#> 35       0.51     0.51 0.6532942
#> 36       0.51     0.51 0.6918628
#> 37       0.51     0.51 0.5587767
#> 38       0.51     0.51 0.5921486
#> 39       0.51     0.51 0.4488923
#> 40       0.51     0.51 0.6391213
#> 41       0.51     0.51 0.7000632
#> 42       0.51     0.51 0.5016235
#> 43       0.51     0.51 0.4730753
#> 44       0.51     0.51 0.6227252
#> 45       0.51     0.51 0.4644939
#> 46       0.51     0.51 0.3279884
#> 47       0.51     0.51 0.4612605
#> 48       0.51     0.51 0.4452160
#> 49       0.51     0.51 0.2132271
#> 50       0.51     0.51 0.5240069
#> 51       0.51     0.51 0.6753944
#> 52       0.51     0.51 0.4761130
#> 53       0.51     0.51 0.2914178
#> 54       0.51     0.51 0.4372740
#> 55       0.51     0.51 0.4657001
#> 56       0.51     0.51 0.6948202
#> 57       0.51     0.51 0.6899856
#> 58       0.51     0.51 0.3577111
#> 59       0.51     0.51 0.7098428
#> 60       0.51     0.51 0.7677141
#> 61       0.51     0.51 0.4549356
#> 62       0.51     0.51 0.6875062
#> 63       0.51     0.51 0.4784716
#> 64       0.51     0.51 0.5075750
#> 65       0.51     0.51 0.6864008
#> 66       0.51     0.51 0.5838673
#> 67       0.51     0.51 0.2718823
#> 68       0.51     0.51 0.4943988
#> 69       0.51     0.51 0.4024763
#> 70       0.51     0.51 0.5009399
#> 71       0.51     0.51 0.4662318
#> 72       0.51     0.51 0.6191566
#> 73       0.51     0.51 0.6424283
#> 74       0.51     0.51 0.6189662
#> 75       0.51     0.51 0.4311113
#> 76       0.51     0.51 0.4797911
#> 77       0.51     0.51 0.5263286
#> 78       0.51     0.51 0.3288324
#> 79       0.51     0.51 0.6114254
#> 80       0.51     0.51 0.7198753
#> 81       0.51     0.51 0.7788448
#> 82       0.51     0.51 0.6955014
#> 83       0.51     0.51 0.6630965
#> 84       0.51     0.51 0.6434546
#> 85       0.51     0.51 0.7085052
#> 86       0.51     0.51 0.4874997
#> 87       0.51     0.51 0.3765264
#> 88       0.51     0.51 0.7927037
#> 89       0.51     0.51 0.3581461
#> 90       0.51     0.51 0.6292601
#> 91       0.51     0.51 0.4099089
#> 92       0.51     0.51 0.4214619
#> 93       0.51     0.51 0.5903802
#> 94       0.51     0.51 0.7699203
#> 95       0.51     0.51 0.3643677
#> 96       0.51     0.51 0.6449606
#> 97       0.51     0.51 0.3220379
#> 98       0.51     0.51 0.5604368
#> 99       0.51     0.51 0.7590660
#> 100      0.51     0.51 0.4607523