Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object

cornet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

"probability", "odds", "log-odds"

...

further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
#> Multiple minima.
predict(net,newx=X)
#>     intercept  gaussian  binomial  combined    switch
#> 1        0.49 0.4739367 0.4636189 0.4478084 0.4636189
#> 2        0.49 0.4750517 0.5315302 0.5011131 0.5315302
#> 3        0.49 0.4405744 0.4920734 0.4412292 0.4920734
#> 4        0.49 0.4848924 0.5804390 0.5478199 0.5804390
#> 5        0.49 0.5100580 0.5136984 0.5199666 0.5136984
#> 6        0.49 0.4512761 0.4292040 0.4015045 0.4292040
#> 7        0.49 0.4378476 0.5351954 0.4723029 0.5351954
#> 8        0.49 0.4784331 0.4520555 0.4430059 0.4520555
#> 9        0.49 0.4831718 0.4963965 0.4815133 0.4963965
#> 10       0.49 0.4466096 0.4560897 0.4183582 0.4560897
#> 11       0.49 0.4646791 0.4175934 0.4040811 0.4175934
#> 12       0.49 0.5070173 0.5447455 0.5410310 0.5447455
#> 13       0.49 0.4583823 0.5547631 0.5042115 0.5547631
#> 14       0.49 0.4293776 0.5618836 0.4864682 0.5618836
#> 15       0.49 0.5312862 0.5404679 0.5600428 0.5404679
#> 16       0.49 0.4736071 0.4726616 0.4544724 0.4726616
#> 17       0.49 0.4243210 0.4061550 0.3629213 0.4061550
#> 18       0.49 0.4154178 0.5561292 0.4723359 0.5561292
#> 19       0.49 0.3959281 0.4373906 0.3691989 0.4373906
#> 20       0.49 0.5225255 0.5099549 0.5286200 0.5099549
#> 21       0.49 0.3779050 0.4348852 0.3584013 0.4348852
#> 22       0.49 0.4336613 0.5105128 0.4500969 0.5105128
#> 23       0.49 0.4081870 0.3608322 0.3173580 0.3608322
#> 24       0.49 0.5142598 0.4216100 0.4529711 0.4216100
#> 25       0.49 0.4795779 0.3962441 0.4010820 0.3962441
#> 26       0.49 0.3793769 0.3994369 0.3317604 0.3994369
#> 27       0.49 0.5501435 0.6290987 0.6445690 0.6290987
#> 28       0.49 0.3851218 0.4206634 0.3507770 0.4206634
#> 29       0.49 0.4376529 0.4722272 0.4236667 0.4722272
#> 30       0.49 0.4165367 0.5203636 0.4455346 0.5203636
#> 31       0.49 0.4980197 0.5376869 0.5271615 0.5376869
#> 32       0.49 0.4527859 0.4697936 0.4340225 0.4697936
#> 33       0.49 0.4499171 0.5256757 0.4746576 0.5256757
#> 34       0.49 0.4455305 0.4749785 0.4320259 0.4749785
#> 35       0.49 0.4239440 0.5250805 0.4542283 0.5250805
#> 36       0.49 0.4414296 0.3848152 0.3593157 0.3848152
#> 37       0.49 0.4562985 0.5176416 0.4738411 0.5176416
#> 38       0.49 0.4662662 0.4235881 0.4101009 0.4235881
#> 39       0.49 0.4764170 0.4795919 0.4623646 0.4795919
#> 40       0.49 0.4921374 0.5187418 0.5070635 0.5187418
#> 41       0.49 0.5170472 0.4229893 0.4566164 0.4229893
#> 42       0.49 0.4637650 0.4493246 0.4277091 0.4493246
#> 43       0.49 0.4630697 0.4432502 0.4224212 0.4432502
#> 44       0.49 0.4169223 0.5179842 0.4439584 0.5179842
#> 45       0.49 0.4914545 0.4651196 0.4651359 0.4651196
#> 46       0.49 0.4419719 0.5244995 0.4673025 0.5244995
#> 47       0.49 0.4212662 0.4565830 0.3996207 0.4565830
#> 48       0.49 0.4475446 0.4318263 0.4004391 0.4318263
#> 49       0.49 0.4368786 0.4545394 0.4094494 0.4545394
#> 50       0.49 0.4854098 0.4925776 0.4806466 0.4925776
#> 51       0.49 0.3710324 0.3549855 0.2939833 0.3549855
#> 52       0.49 0.4789286 0.5790026 0.5412097 0.5790026
#> 53       0.49 0.4784493 0.6057435 0.5613605 0.6057435
#> 54       0.49 0.4590848 0.4696856 0.4393076 0.4696856
#> 55       0.49 0.4327928 0.4507130 0.4033979 0.4507130
#> 56       0.49 0.4632548 0.4545810 0.4313084 0.4545810
#> 57       0.49 0.4842948 0.5282670 0.5070933 0.5282670
#> 58       0.49 0.4498660 0.4932365 0.4496370 0.4932365
#> 59       0.49 0.4760681 0.4622974 0.4487297 0.4622974
#> 60       0.49 0.5024997 0.6041833 0.5825656 0.6041833
#> 61       0.49 0.4949405 0.4806575 0.4803624 0.4806575
#> 62       0.49 0.4532852 0.4503085 0.4194391 0.4503085
#> 63       0.49 0.4356939 0.4805819 0.4285932 0.4805819
#> 64       0.49 0.4610748 0.4967189 0.4618489 0.4967189
#> 65       0.49 0.4702076 0.4440982 0.4294115 0.4440982
#> 66       0.49 0.3459168 0.4184492 0.3343943 0.4184492
#> 67       0.49 0.5106581 0.5165648 0.5227336 0.5165648
#> 68       0.49 0.4563607 0.5469444 0.4964573 0.5469444
#> 69       0.49 0.4466603 0.4095836 0.3825898 0.4095836
#> 70       0.49 0.4396995 0.4725668 0.4255218 0.4725668
#> 71       0.49 0.4659226 0.5104406 0.4766727 0.5104406
#> 72       0.49 0.4158239 0.4478045 0.3891930 0.4478045
#> 73       0.49 0.3550303 0.4334065 0.3486118 0.4334065
#> 74       0.49 0.5269767 0.4756933 0.5062896 0.4756933
#> 75       0.49 0.4617527 0.5578567 0.5095160 0.5578567
#> 76       0.49 0.4982556 0.5917150 0.5689844 0.5917150
#> 77       0.49 0.5192314 0.5839596 0.5825794 0.5839596
#> 78       0.49 0.4977590 0.5315953 0.5222264 0.5315953
#> 79       0.49 0.4661216 0.4405829 0.4230587 0.4405829
#> 80       0.49 0.4908152 0.5605508 0.5380206 0.5605508
#> 81       0.49 0.4665530 0.4835217 0.4565042 0.4835217
#> 82       0.49 0.3628216 0.4395791 0.3559995 0.4395791
#> 83       0.49 0.4708090 0.4642303 0.4454536 0.4642303
#> 84       0.49 0.3985524 0.4841634 0.4066623 0.4841634
#> 85       0.49 0.4875089 0.4239547 0.4297583 0.4239547
#> 86       0.49 0.4505878 0.5056177 0.4597697 0.5056177
#> 87       0.49 0.4861137 0.5674455 0.5389487 0.5674455
#> 88       0.49 0.4246604 0.4697539 0.4121323 0.4697539
#> 89       0.49 0.4058184 0.5211569 0.4393664 0.5211569
#> 90       0.49 0.4637595 0.4934383 0.4616719 0.4934383
#> 91       0.49 0.5631325 0.5503503 0.5943243 0.5503503
#> 92       0.49 0.5127034 0.5543157 0.5537075 0.5543157
#> 93       0.49 0.4787031 0.4778743 0.4631340 0.4778743
#> 94       0.49 0.5039298 0.5701190 0.5576768 0.5701190
#> 95       0.49 0.5449092 0.4816308 0.5266270 0.4816308
#> 96       0.49 0.4661828 0.5189345 0.4834436 0.5189345
#> 97       0.49 0.4964894 0.5885897 0.5649218 0.5885897
#> 98       0.49 0.5062561 0.5607779 0.5526635 0.5607779
#> 99       0.49 0.5159205 0.4905809 0.5076191 0.4905809
#> 100      0.49 0.5249085 0.4669833 0.4977061 0.4669833