Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object
cornet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
"probability", "odds", "log-odds"

...
further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)

#>     intercept  binomial  combined
#> 1        0.49 0.5268111 0.5075069
#> 2        0.49 0.5701215 0.5356587
#> 3        0.49 0.4959983 0.4874786
#> 4        0.49 0.4344885 0.4474973
#> 5        0.49 0.5051122 0.4934026
#> 6        0.49 0.4473818 0.4558779
#> 7        0.49 0.4593130 0.4636332
#> 8        0.49 0.4422506 0.4525426
#> 9        0.49 0.4825321 0.4787256
#> 10       0.49 0.5215738 0.5041027
#> 11       0.49 0.4616143 0.4651290
#> 12       0.49 0.5200346 0.5031022
#> 13       0.49 0.5109024 0.4971663
#> 14       0.49 0.5443871 0.5189314
#> 15       0.49 0.4251630 0.4414357
#> 16       0.49 0.4702473 0.4707405
#> 17       0.49 0.4545353 0.4605276
#> 18       0.49 0.5101508 0.4966777
#> 19       0.49 0.4414225 0.4520043
#> 20       0.49 0.4672553 0.4687957
#> 21       0.49 0.5466004 0.5203700
#> 22       0.49 0.4875851 0.4820100
#> 23       0.49 0.5158311 0.5003699
#> 24       0.49 0.4824406 0.4786661
#> 25       0.49 0.4999042 0.4900174
#> 26       0.49 0.5379723 0.5147617
#> 27       0.49 0.4749272 0.4737824
#> 28       0.49 0.5392102 0.5155663
#> 29       0.49 0.4773595 0.4753634
#> 30       0.49 0.4928284 0.4854181
#> 31       0.49 0.5008112 0.4906070
#> 32       0.49 0.5213078 0.5039298
#> 33       0.49 0.4955747 0.4872033
#> 34       0.49 0.4683778 0.4695253
#> 35       0.49 0.4373248 0.4493408
#> 36       0.49 0.5029416 0.4919918
#> 37       0.49 0.4518527 0.4587840
#> 38       0.49 0.4951047 0.4868978
#> 39       0.49 0.5063450 0.4942040
#> 40       0.49 0.4958585 0.4873877
#> 41       0.49 0.5082019 0.4954109
#> 42       0.49 0.4837213 0.4794986
#> 43       0.49 0.5001347 0.4901672
#> 44       0.49 0.4991447 0.4895238
#> 45       0.49 0.4649256 0.4672814
#> 46       0.49 0.5066679 0.4944139
#> 47       0.49 0.5084433 0.4955678
#> 48       0.49 0.5339739 0.5121628
#> 49       0.49 0.4750754 0.4738787
#> 50       0.49 0.5003858 0.4903305
#> 51       0.49 0.5106505 0.4970025
#> 52       0.49 0.4560670 0.4615233
#> 53       0.49 0.5516260 0.5236366
#> 54       0.49 0.5185769 0.5021547
#> 55       0.49 0.4769370 0.4750888
#> 56       0.49 0.4916435 0.4846480
#> 57       0.49 0.4479826 0.4562684
#> 58       0.49 0.5485780 0.5216554
#> 59       0.49 0.4577390 0.4626100
#> 60       0.49 0.4926756 0.4853189
#> 61       0.49 0.4791468 0.4765251
#> 62       0.49 0.3889889 0.4179225
#> 63       0.49 0.4958569 0.4873867
#> 64       0.49 0.4889662 0.4829078
#> 65       0.49 0.4723737 0.4721226
#> 66       0.49 0.4833173 0.4792360
#> 67       0.49 0.4545089 0.4605105
#> 68       0.49 0.4765813 0.4748576
#> 69       0.49 0.5150237 0.4998451
#> 70       0.49 0.4603223 0.4642892
#> 71       0.49 0.4798057 0.4769534
#> 72       0.49 0.5324028 0.5111415
#> 73       0.49 0.5266367 0.5073936
#> 74       0.49 0.4826416 0.4787967
#> 75       0.49 0.5083714 0.4955211
#> 76       0.49 0.4655858 0.4677105
#> 77       0.49 0.4716670 0.4716632
#> 78       0.49 0.4684682 0.4695840
#> 79       0.49 0.4842055 0.4798133
#> 80       0.49 0.4967722 0.4879817
#> 81       0.49 0.4599805 0.4640670
#> 82       0.49 0.5344043 0.5124425
#> 83       0.49 0.4721848 0.4719998
#> 84       0.49 0.5068752 0.4945486
#> 85       0.49 0.5434044 0.5182926
#> 86       0.49 0.4574339 0.4624118
#> 87       0.49 0.4531499 0.4596272
#> 88       0.49 0.4999355 0.4900378
#> 89       0.49 0.4948965 0.4867624
#> 90       0.49 0.4414958 0.4520520
#> 91       0.49 0.4348078 0.4477048
#> 92       0.49 0.4816584 0.4781577
#> 93       0.49 0.4318491 0.4457816
#> 94       0.49 0.5345482 0.5125360
#> 95       0.49 0.5543119 0.5253825
#> 96       0.49 0.5462319 0.5201305
#> 97       0.49 0.4762285 0.4746283
#> 98       0.49 0.5142520 0.4993435
#> 99       0.49 0.4510697 0.4582750
#> 100      0.49 0.5250353 0.5063526