Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object
cornet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
"probability", "odds", "log-odds"

...
further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)

#>      gaussian  binomial  combined
#> 1   0.4963616 0.4310056 0.4396183
#> 2   0.4611006 0.5952511 0.5626617
#> 3   0.5369638 0.3770472 0.4129054
#> 4   0.4622992 0.5276183 0.5057223
#> 5   0.4276312 0.4233198 0.4019431
#> 6   0.4885601 0.6339410 0.6083962
#> 7   0.5086451 0.4041088 0.4226160
#> 8   0.4473326 0.6283423 0.5846050
#> 9   0.4397399 0.5043428 0.4759003
#> 10  0.5103900 0.6570764 0.6384691
#> 11  0.4607411 0.5804291 0.5498988
#> 12  0.5111168 0.5449067 0.5434706
#> 13  0.4319505 0.4751688 0.4478020
#> 14  0.5243871 0.5637356 0.5657498
#> 15  0.5349596 0.5942640 0.5966180
#> 16  0.4781994 0.5612015 0.5416626
#> 17  0.4829494 0.4590095 0.4570422
#> 18  0.5624767 0.6263197 0.6361117
#> 19  0.5223702 0.6722374 0.6570285
#> 20  0.4418833 0.4611428 0.4401050
#> 21  0.5133652 0.4992129 0.5056995
#> 22  0.4724334 0.5492985 0.5288408
#> 23  0.5079102 0.5514791 0.5475306
#> 24  0.5686275 0.4916058 0.5241973
#> 25  0.4544229 0.4508476 0.4368912
#> 26  0.5424608 0.4356977 0.4652656
#> 27  0.5655013 0.4296630 0.4702359
#> 28  0.5108477 0.4836688 0.4912904
#> 29  0.4837504 0.5521184 0.5365640
#> 30  0.5440719 0.5019949 0.5223468
#> 31  0.4563561 0.6118615 0.5746243
#> 32  0.4208328 0.4329792 0.4074068
#> 33  0.5339492 0.5766883 0.5812125
#> 34  0.4521809 0.5208268 0.4953685
#> 35  0.5189346 0.5302303 0.5347026
#> 36  0.5197830 0.6094299 0.6024229
#> 37  0.4719330 0.5574237 0.5355136
#> 38  0.4727123 0.4579008 0.4512831
#> 39  0.4700989 0.4824937 0.4709686
#> 40  0.5273016 0.5004851 0.5133514
#> 41  0.4627463 0.4428144 0.4338440
#> 42  0.5770216 0.6184816 0.6354769
#> 43  0.4806800 0.5049115 0.4949859
#> 44  0.5002068 0.5803177 0.5683688
#> 45  0.4028669 0.3708666 0.3477769
#> 46  0.5392254 0.5473850 0.5587278
#> 47  0.4456822 0.4674736 0.4471412
#> 48  0.5393637 0.4263724 0.4559303
#> 49  0.4743298 0.5253378 0.5093611
#> 50  0.5447890 0.4760307 0.5006007
#> 51  0.4184321 0.3563414 0.3413150
#> 52  0.4528447 0.5171037 0.4925008
#> 53  0.4869225 0.6447837 0.6168342
#> 54  0.4605769 0.4258334 0.4184175
#> 55  0.4861294 0.5681202 0.5512937
#> 56  0.5023638 0.6054661 0.5907743
#> 57  0.4865707 0.6328550 0.6065278
#> 58  0.4791707 0.3831914 0.3908117
#> 59  0.5186321 0.4578319 0.4730211
#> 60  0.4526632 0.4420662 0.4286376
#> 61  0.5043652 0.7479601 0.7128492
#> 62  0.5433787 0.6066142 0.6109601
#> 63  0.4785587 0.7815771 0.7291511
#> 64  0.5080174 0.5953164 0.5848434
#> 65  0.5111599 0.4954458 0.5014494
#> 66  0.4614964 0.4440986 0.4343630
#> 67  0.4555427 0.5306671 0.5052420
#> 68  0.5643650 0.6419300 0.6501828
#> 69  0.4518096 0.4953327 0.4735327
#> 70  0.4610520 0.5773794 0.5474485
#> 71  0.4268002 0.3904596 0.3736718
#> 72  0.4266856 0.4599569 0.4326976
#> 73  0.4822425 0.3816143 0.3909217
#> 74  0.5134585 0.5869770 0.5803433
#> 75  0.4212360 0.4204230 0.3968947
#> 76  0.4422157 0.5112867 0.4828712
#> 77  0.5371228 0.4751262 0.4963455
#> 78  0.5509981 0.4631255 0.4924052
#> 79  0.4589552 0.5359392 0.5112685
#> 80  0.4891379 0.4531376 0.4549882
#> 81  0.5881646 0.6004634 0.6245221
#> 82  0.5166605 0.6205038 0.6103595
#> 83  0.4891233 0.5553168 0.5418335
#> 84  0.5444372 0.4857899 0.5087375
#> 85  0.5381042 0.7150333 0.7007163
#> 86  0.4919651 0.4907524 0.4883067
#> 87  0.4798224 0.3569478 0.3688119
#> 88  0.4016109 0.3811948 0.3561027
#> 89  0.4921590 0.4733724 0.4736261
#> 90  0.4816429 0.4383202 0.4388382
#> 91  0.5456118 0.4383354 0.4689300
#> 92  0.4318745 0.6089215 0.5614602
#> 93  0.4867017 0.4898072 0.4849993
#> 94  0.5001538 0.6428465 0.6214929
#> 95  0.5032944 0.4369885 0.4480126
#> 96  0.5642937 0.4582474 0.4940225
#> 97  0.5537003 0.5722639 0.5863646
#> 98  0.5301367 0.5687940 0.5727356
#> 99  0.4650822 0.5257808 0.5054394
#> 100 0.5238035 0.6059937 0.6013951