Predict binary outcome

Predicts the binary outcome with linear, logistic, and combined regression.

# S3 method for cornet
predict(object, newx, type = "probability", ...)

Arguments

object
cornet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
"probability", "odds", "log-odds"

...
further arguments (not applicable)

Details

For linear regression, this function tentatively transforms the predicted values to predicted probabilities, using a Gaussian distribution with a fixed mean (threshold) and a fixed variance (estimated variance of the numeric outcome).

Examples

n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
predict(net,newx=X)

#>      gaussian binomial  combined
#> 1   0.4735457     0.57 0.4879335
#> 2   0.4049645     0.57 0.4029510
#> 3   0.6111345     0.57 0.6593895
#> 4   0.5172448     0.57 0.5437339
#> 5   0.4998924     0.57 0.5215625
#> 6   0.5821378     0.57 0.6249348
#> 7   0.4125119     0.57 0.4119956
#> 8   0.5675041     0.57 0.6070186
#> 9   0.5113804     0.57 0.5362471
#> 10  0.3712594     0.57 0.3639163
#> 11  0.4970856     0.57 0.5179735
#> 12  0.5309086     0.57 0.5611262
#> 13  0.5738117     0.57 0.6147785
#> 14  0.6318453     0.57 0.6829535
#> 15  0.4061513     0.57 0.4043666
#> 16  0.6953709     0.57 0.7481822
#> 17  0.6062483     0.57 0.6536945
#> 18  0.6201689     0.57 0.6697869
#> 19  0.5539592     0.57 0.5901863
#> 20  0.5846857     0.57 0.6280215
#> 21  0.6176490     0.57 0.6669046
#> 22  0.6124788     0.57 0.6609476
#> 23  0.4901677     0.57 0.5091306
#> 24  0.5269040     0.57 0.5560383
#> 25  0.5934541     0.57 0.6385621
#> 26  0.4932075     0.57 0.5130156
#> 27  0.6542923     0.57 0.7073161
#> 28  0.6275240     0.57 0.6781178
#> 29  0.6041358     0.57 0.6512175
#> 30  0.5179353     0.57 0.5446148
#> 31  0.5688653     0.57 0.6086977
#> 32  0.4893587     0.57 0.5080970
#> 33  0.5408113     0.57 0.5736636
#> 34  0.4488653     0.57 0.4567495
#> 35  0.6008167     0.57 0.6473081
#> 36  0.5090293     0.57 0.5332433
#> 37  0.6343333     0.57 0.6857171
#> 38  0.4454156     0.57 0.4524322
#> 39  0.4479874     0.57 0.4556496
#> 40  0.5860962     0.57 0.6297257
#> 41  0.5849100     0.57 0.6282927
#> 42  0.6536787     0.57 0.7066679
#> 43  0.4955070     0.57 0.5159552
#> 44  0.5541269     0.57 0.5903960
#> 45  0.5939806     0.57 0.6391907
#> 46  0.4961191     0.57 0.5167377
#> 47  0.4210566     0.57 0.4223495
#> 48  0.4969161     0.57 0.5177568
#> 49  0.5515815     0.57 0.5872104
#> 50  0.5575421     0.57 0.5946593
#> 51  0.5328922     0.57 0.5636429
#> 52  0.5087409     0.57 0.5328746
#> 53  0.5541515     0.57 0.5904267
#> 54  0.5402391     0.57 0.5729411
#> 55  0.5314805     0.57 0.5618520
#> 56  0.5638982     0.57 0.6025589
#> 57  0.5051317     0.57 0.5282614
#> 58  0.5826846     0.57 0.6255981
#> 59  0.5404302     0.57 0.5731824
#> 60  0.5385386     0.57 0.5707925
#> 61  0.5863818     0.57 0.6300704
#> 62  0.5360255     0.57 0.5676131
#> 63  0.4930078     0.57 0.5127602
#> 64  0.4361175     0.57 0.4408605
#> 65  0.4844202     0.57 0.5017907
#> 66  0.5174933     0.57 0.5440510
#> 67  0.4679033     0.57 0.4807661
#> 68  0.5251721     0.57 0.5538352
#> 69  0.5904760     0.57 0.6349968
#> 70  0.5627922     0.57 0.6011878
#> 71  0.6334061     0.57 0.6846889
#> 72  0.4613113     0.57 0.4724178
#> 73  0.4426430     0.57 0.4489714
#> 74  0.5973167     0.57 0.6431629
#> 75  0.5496140     0.57 0.5847437
#> 76  0.4575957     0.57 0.4677263
#> 77  0.5370906     0.57 0.5689612
#> 78  0.5481692     0.57 0.5829299
#> 79  0.6031529     0.57 0.6500620
#> 80  0.4673421     0.57 0.4800542
#> 81  0.4361413     0.57 0.4408901
#> 82  0.5496858     0.57 0.5848338
#> 83  0.5758016     0.57 0.6172151
#> 84  0.6181958     0.57 0.6675312
#> 85  0.5451207     0.57 0.5790964
#> 86  0.5495116     0.57 0.5846151
#> 87  0.4996389     0.57 0.5212382
#> 88  0.5227819     0.57 0.5507923
#> 89  0.4862134     0.57 0.5040798
#> 90  0.5330534     0.57 0.5638473
#> 91  0.5247780     0.57 0.5533337
#> 92  0.4775941     0.57 0.4930866
#> 93  0.6090519     0.57 0.6569681
#> 94  0.4799998     0.57 0.4961522
#> 95  0.6142294     0.57 0.6629711
#> 96  0.4046466     0.57 0.4025722
#> 97  0.4782808     0.57 0.4939614
#> 98  0.5492885     0.57 0.5843351
#> 99  0.4894120     0.57 0.5081652
#> 100 0.5436622     0.57 0.5772595